% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{clean}
\alias{clean}
\title{clean}
\usage{
clean(data)
}
\arguments{
\item{data}{the csv from a vcf file

delete the monomorphic loci of a vcf file imported as a CSV#'}
}
\description{
clean
}
\author{
JAUNATRE Maxime
}
