% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{durand_freq_D}
\alias{durand_freq_D}
\title{ABBA-BABA test}
\usage{
durand_freq_D(file, P1, P2, P3, Root, n)
}
\arguments{
\item{file}{a list containning an element named .csv (a character chain with the location of a .csv file) and a vector of factor .pop containing the assignation to populations for each individual. usually made by the dataset function}

\item{P1, P2, P3, Root}{a character chain for the populations to compare in this test}

\item{n}{number of bootstrap.}
}
\value{
a list with a vector (all the D value, first one is without bootstrap) and a value for the number of informativ loci
}
\description{
compute the durand's D for admixture test between different populations
}
\author{
JAUNATRE Maxime
}
