% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{files}
\alias{files}
\title{files}
\usage{
files(obj, name, ind, pop, head = NULL)
}
\arguments{
\item{obj}{the data frame originaly made from a vcf file}

\item{name}{the name of the file to output}

\item{ind}{the list of individuals}

\item{pop}{the list of population assignations}

\item{head}{a txt file with the vcf head to take from}
}
\description{
create files for the population genetic analysis
.vcf
.str (structure)
.geno (genotype for LEA)
.snp (for DIYABC)
}
\author{
JAUNATRE Maxime
}
