% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{\%.in\%}
\alias{\%.in\%}
\title{not in}
\usage{
x \%.in\% y
}
\arguments{
\item{x}{one object}

\item{y}{second object

The negate of the %in% function :
match returns a vector of the positions of (first) matches of its first argument NOT in its second.}
}
\description{
not in
}
\author{
JAUNATRE Maxime from (https://stackoverflow.com/questions/5831794/opposite-of-in)
}
