% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{objvcf2geno}
\alias{objvcf2geno}
\title{objvcf2geno}
\usage{
objvcf2geno(data, name)
}
\arguments{
\item{name}{name for save}

\item{obj}{name of the data frame to save}
}
\description{
save a dataframe as a vcf file, concatening the head of a vcf
}
\author{
JAUNATRE Maxime
}
