% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{plot.raster}
\alias{plot.raster}
\title{plot a raster}
\usage{
plot.raster(rast, ext, T.ext = F, line = F)
}
\arguments{
\item{rast}{the raster file}

\item{ext}{the extent of this raster}

\item{T.ext}{T if you want to crop with an extent}

\item{line}{T if you want the water limit}
}
\description{
plot a raster
}
\author{
JAUNATRE Maxime
}
