% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{Pop}
\alias{Pop}
\title{Pop}
\usage{
Pop(K, files, ID)
}
\arguments{
\item{ID}{individual tags

barplot the structure of the geno file}

\item{k}{the K to test}

\item{the}{file use in cross entropy (.geno)}
}
\description{
str barplot
}
\author{
JAUNATRE Maxime
}
