% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{rare}
\alias{rare}
\title{rare}
\usage{
rare(data, rare = 0, quiet = T, r = F, p = F)
}
\arguments{
\item{data}{the csv from a vcf file}

\item{rare}{percentage of presence needed for keeping the allele in the set

delete the alleles with less than a "rare" percentage of presence in the dataset}
}
\description{
rare
}
\author{
JAUNATRE Maxime
}
