% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{subset_ord_pop}
\alias{subset_ord_pop}
\title{reorder and subset the population assignement}
\usage{
subset_ord_pop(population, sub)
}
\arguments{
\item{population}{a dataframe with two column ind and pop, row is an individual and its population assignement}

\item{sub}{a vector with the individuals to subset

reorder and subset the population assignement

return a data.frame}
}
\description{
reorder and subset the population assignement
}
\author{
JAUNATRE Maxime
}
