% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{subset_reorder}
\alias{subset_reorder}
\title{subset the main information with the list of individual and clean the dataset for monomorphic loci}
\usage{
subset_reorder(data, list)
}
\arguments{
\item{data}{a data frame from a vcf file (without the hashtag lignes)}

\item{list}{vector of individual to maintain in this subset (individual ID must be the same as the colnames in the dataframe)}
}
\description{
subset the main information with the list of individual and clean the dataset for monomorphic loci
}
\author{
JAUNATRE Maxime
}
