% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{tablobj2vcf}
\alias{tablobj2vcf}
\title{tablobj2vcf}
\usage{
tablobj2vcf(obj, name, head, vcf)
}
\arguments{
\item{obj}{name of the data frame to save}

\item{name}{name for save}

\item{head}{the txt file where is stock the head of the vcf}

\item{vcf}{name of the final file}
}
\description{
save a dataframe as a vcf file, concatening the head of a vcf
}
\author{
JAUNATRE Maxime
}
