% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{tri}
\alias{tri}
\title{clean}
\usage{
tri(data, n.r = 0, n.c = 0, quiet = F, r = F, p = F)
}
\arguments{
\item{data}{the csv from a vcf file}

\item{n.r}{percentage of row containing information (1-missing data)}

\item{quiet}{if the function print information at the end about percentage of data discarded}

\item{r}{if the function return the data or the resume}

\item{p}{if the function return the resume only

delete in a first time the loci with a certain percentage of missing data, and in a second time individuals with too many missing data}

\item{c.r}{idem but for individuals}
}
\description{
clean
}
\author{
JAUNATRE Maxime
}
