% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{clean}
\alias{clean}
\title{clean}
\usage{
clean(data)
}
\arguments{
\item{data}{a data frame from a vcf file (without the hashtag lignes)}
}
\value{
a data frame
}
\description{
delete the monomorphic loci of a vcf file imported as a CSV
}
\author{
JAUNATRE Maxime
}
