% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{dataset}
\alias{dataset}
\title{create de dataset files}
\usage{
dataset(ind, popfile, entryfile, name, rare, qual, missLoci, missInd, LD)
}
\arguments{
\item{ind}{vector of individuals}

\item{popfile}{csv file with population assignement}

\item{entryfile}{csv file originally a vcf}

\item{name}{basic name and location of the files to (character chain)}

\item{rare}{numeric for the percentage of presence minimal for an allele to be retain (ex 0.05)}

\item{qual}{numeric for the quality of a loci to be retain (ex 20)}

\item{missLoci}{numeric for the percentage of missing data of an allele to be retain (ex 0.8)}

\item{missInd}{numeric for the percentage of missing data of an individual to be retain (ex 0.8)}

\item{LD}{numeric, minimal distance between two loci on a contig (ex 1e4)

create all thes files for population genetic

return a list with all the files created for this analysis}
}
\description{
create de dataset files
}
\author{
JAUNATRE Maxime
}
