% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{save2vcf}
\alias{save2vcf}
\title{save as a vcf file but need to be a vcf troncated at the origine
hide because tablobj2vcf is way better}
\usage{
save2vcf(csv)
}
\arguments{
\item{csv}{the dataframe with the information to save inside a vcf file}
}
\description{
save as a vcf file but need to be a vcf troncated at the origine
hide because tablobj2vcf is way better
}
\author{
JAUNATRE Maxime
}
