% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{spider}
\alias{spider}
\title{spider}
\usage{
spider(input, inFORM, output, outFORM)
}
\arguments{
\item{input}{the name of the entry file}

\item{inFORM}{the type of entry file}

\item{output}{the name of the exit file}

\item{outFORM}{the type of exit file}
}
\description{
push file from a type to another using pgdspider.
For now, just working with VCF to structure, because it need a .spid done for the software
}
\author{
JAUNATRE Maxime
}
