% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{subset_reorder}
\alias{subset_reorder}
\title{subset and reorder}
\usage{
subset_reorder(data, list)
}
\arguments{
\item{data}{the csv from a vcf file}

\item{list}{list of individual to maintain in this subset

subset the main information with the list of individual and clean the dataset for monomorphic loci}
}
\description{
subset and reorder
}
\author{
JAUNATRE Maxime
}
