% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{vcf2csv}
\alias{vcf2csv}
\title{vcf2csv}
\usage{
vcf2csv(vcf, head, csv)
}
\arguments{
\item{vcf}{the original vcf file}

\item{head}{the txt file where to stock the head of the vcf}

\item{csv}{the name of the final csv}
}
\description{
save as a csv file a vcf file, allowing to apply treshold on the data
}
\author{
JAUNATRE Maxime
}
